fix(cli): route pet edits through bearer API
